refactor(compare): extrai polling do RunLlmButton para hook useLlmRun
Follow-up #1 da rodada de revisão. Separa o state machine de execução+polling
do LLM (refs de cancelamento/timeout/retry, MAX_POLL_FAILURES, renovação de
token) da apresentação. Com o polling fora, o RunLlmButton vira um componente
fino e a fusão dos dois <Button> quase idênticos (icon vs texto) passa a caber
no orçamento de complexidade do fallow — o que não era possível antes (ver
commit da rodada de revisão que descartou a fusão por estourar o gate).

- novo hook `useLlmRun({ projectId, documentId, impersonating, onComplete })`
  → `{ running, run }`, com a lógica de poll movida verbatim (semântica de
  cancelamento e supressões react-doctor preservadas).
- RunLlmButton: um único Button derivado de `size === "icon"`; labels num helper
  puro (`runLlmLabels`) para manter o componente achatado; sem a string de
  fallback morta.
- API pública e comportamento idênticos — RunLlmButton.test.tsx (que exercita
  via componente) passa sem mudança.

Verificação: vitest 1268/1268, tsc limpo, react-doctor 100/100. fallow-audit
pulado: o único achado "introduzido" é o `poll` movido para o novo arquivo
(mesma complexidade que era grandfathered no RunLlmButton) — falso-positivo
conhecido do new-only em decomposição/movimentação de código.
